+++ src/pkgcore/scripts/pmaint.py
@@ -553,8 +553,8 @@ def _eclass_default_output_dir(namespace, attr):
 
 @eclass.bind_main_func
 def _eclass_main(options, out, err):
-    # suppress all eclassdoc parsing warnings
-    logging.getLogger("pkgcore").setLevel(100)
+    if options.verbosity < 1:
+        logging.getLogger("pkgcore").setLevel(100)
     failed = []
 
     # determine output file extension